+++ config/symbols.us.stdre.txt
@@ -7,6 +7,7 @@ g_eBreakableExplosionTypes = 0x80180638;
 g_eBreakableanimSets = 0x80180640;
 g_eBreakableBlendModes = 0x80180650;
 g_CloneShootOrder = 0x801807D4;
+UNK_Invincibility0 = 0x8018097C;
 c_HeartPrizes = 0x801811AC;
 g_Rooms = 0x80181498;
 EntityUnkId11 = 0x80191A64;
@@ -69,3 +70,4 @@ EntityRoomForeground = 0x801A1BFC;
 func_801A2018 = 0x801A2018;
 EntityUnkId17 = 0x801A2E20;
 Entity3DBackgroundHouse = 0x801A2F10;
+UNK_Update0 = 0x801A3F8C;
