+++ config/symbols.us.stwrp.txt
@@ -10,6 +10,7 @@ g_eBreakableHitboxes = 0x801805D8;
 g_eBreakableExplosionTypes = 0x801805E0;
 g_eBreakableanimSets = 0x801805E8;
 g_eBreakableBlendModes = 0x801805F8;
+UNK_Invincibility0 = 0x80180690;
 c_GoldPrizes = 0x80180E30;
 c_HeartPrizes = 0x80180EC0;
 g_Rooms = 0x801811AC;
@@ -110,3 +111,4 @@ EntitySoulStealOrb = 0x80192610;
 EntityEnemyBlood = 0x801929A4;
 EntityRoomForeground = 0x80192E54;
 func_80193270 = 0x80193270;
+UNK_Update0 = 0x80194728;
