+++ src/st/cen/D600.c
@@ -898,10 +898,7 @@ void EntityMovingElevator(Entity* self) {
     }
 }
 
-s32 Random(void) {
-    g_randomNext = (g_randomNext * 0x01010101) + 1;
-    return g_randomNext >> 0x18;
-}
+#include "../random.h"
 
 INCLUDE_ASM("asm/us/st/cen/nonmatchings/D600", func_80190E7C);
