+++ src/st/dre/1C7DC.c
@@ -40,7 +40,7 @@ void EntityEquipItemDrop(Entity* self) {
 
         if (!(collider.effects & EFFECT_NOTHROUGH_PLUS)) {
             for (index = 0; index < 32; index++) {
-                if (D_801A3F8C[index] == 0) {
+                if (UNK_Update0[index] == 0) {
                     break;
                 }
             }
@@ -64,7 +64,7 @@ void EntityEquipItemDrop(Entity* self) {
 
             self->flags |= FLAG_HAS_PRIMS;
             self->primIndex = primIndex;
-            D_801A3F8C[index] = 0x1E0;
+            UNK_Update0[index] = 0x1E0;
             self->ext.generic.unk8C.modeU16.unk0 = index;
 
             if (itemId < NUM_HAND_ITEMS) {
@@ -139,7 +139,7 @@ void EntityEquipItemDrop(Entity* self) {
                 self->step++;
             }
         } else {
-            D_801A3F8C[self->ext.generic.unk8C.modeS16.unk0] = 0x10;
+            UNK_Update0[self->ext.generic.unk8C.modeS16.unk0] = 0x10;
         }
         break;
