+++ src/st/dre/dre.h
@@ -92,7 +92,7 @@ extern u8 D_80180780[]; // Animation
 extern const u8 D_80180794[];
 extern s32 D_801807F8[];
 extern u8 D_80180830[];
-extern u16 D_8018097C[];
+extern u16 UNK_Invincibility0[];
 extern s16 D_80180D80[];
 extern u16 D_80180470[];
 extern u16 D_801804AC[];
@@ -148,4 +148,4 @@ extern s32 D_801A3F84;
 extern s16 D_801A3F14;
 extern s16 D_801A3F16;
 extern s32 D_801A3F18;
-extern u16 D_801A3F8C[];
+extern u16 UNK_Update0[];
