+++ src/st/mad/D8C8.c
@@ -151,10 +151,7 @@ void EntityBreakable(Entity* entity) {
     }
 }
 
-s32 Random(void) {
-    g_randomNext = (g_randomNext * 0x01010101) + 1;
-    return g_randomNext >> 0x18;
-}
+#include "../random.h"
 
 void Update(void) {
     s16 i;
