+++ src/st/np3/390BC.c
@@ -1,9 +1,6 @@
 #include "np3.h"
 
-s32 Random(void) {
-    g_randomNext = (g_randomNext * 0x01010101) + 1;
-    return g_randomNext >> 0x18;
-}
+#include "../random.h"
 
 void Update(void) {
     s16 i;
