+++ src/st/random.h
@@ -1,0 +1,4 @@
+s32 Random(void) {
+    g_randomNext = (g_randomNext * 0x01010101) + 1;
+    return g_randomNext >> 0x18;
+}
