+++ src/st/wrp/6FD0.c
@@ -464,7 +464,7 @@ WarpCoord D_8018065C[] = {
 // *** entity declaration end ***
 
 // *** Unknown stuff start ***
-u16 D_80180690[] = {
+u16 UNK_Invincibility0[] = {
     /* 690 */ 0x8163,
     /* 692 */ 0x8164,
     /* 694 */ 0x8166,
@@ -1190,7 +1190,7 @@ extern LayoutEntity* D_80193AB0;
 extern u16* D_80193AB4;
 extern u8 D_80193AB8;
 extern u8 D_80193ABC;
-extern u16 D_80194728[];
+extern u16 UNK_Update0[];
 // *** bss? section end ***
 
 void func_80186FD0(Entity* arg0) {
@@ -1672,93 +1672,9 @@ void EntityWarpSmallRocks(Entity* entity) {
 }
 #endif
 
-s32 Random(void) {
-    g_randomNext = (g_randomNext * 0x01010101) + 1;
-    return g_randomNext >> 0x18;
-}
+#include "../random.h"
 
-void Update(void) {
-    s16 i;
-    Entity* entity;
-    s32* unk;
-
-    for (i = 0; i < 0x20; i++) {
-        if (D_80194728[i]) {
-            D_80194728[i]--;
-        }
-    }
-
-    unk = &D_80097410;
-    if (*unk) {
-        if (!--*unk) {
-            g_api.FreePrimitives(D_80097414);
-        }
-    }
-
-    for (entity = &g_Entities[STAGE_ENTITY_START];
-         entity < &g_Entities[TOTAL_ENTITY_COUNT]; entity++) {
-        if (!entity->pfnUpdate)
-            continue;
-
-        if (entity->step) {
-            s32 unk34 = entity->flags;
-            if (unk34 & FLAG_DESTROY_IF_OUT_OF_CAMERA) {
-                s16 posX = i = entity->posX.i.hi;
-                s16 posY = entity->posY.i.hi;
-                if (unk34 & FLAG_DESTROY_IF_BARELY_OUT_OF_CAMERA) {
-                    if ((u16)(posX + 64) > 384 || (u16)(posY + 64) > 352) {
-                        DestroyEntity(entity);
-                        continue;
-                    }
-                } else {
-                    if ((u16)(posX + 128) > 512 || (u16)(posY + 128) > 480) {
-                        DestroyEntity(entity);
-                        continue;
-                    }
-                }
-            }
-
-            if ((unk34 & 0x02000000)) {
-                s16 posY = entity->posY.i.hi + g_Camera.posY.i.hi;
-                s16 test = (LOHU(g_CurrentRoom.vSize) * 256) + 128;
-                if (posY > test) {
-                    DestroyEntity(entity);
-                    continue;
-                }
-            }
-
-            if (unk34 & 0xF) {
-                entity->palette =
-                    D_80180690[(entity->nFramesInvincibility << 1) |
-                               (unk34 & 1)];
-                entity->flags--;
-                if ((entity->flags & 0xF) == 0) {
-                    entity->palette = entity->unk6A;
-                    entity->unk6A = 0;
-                }
-            }
-
-            if (!(unk34 & 0x20000000) || (unk34 & 0x10000000) ||
-                ((u16)(entity->posX.i.hi + 64) <= 384) &&
-                    ((u16)(entity->posY.i.hi + 64) <= 352)) {
-                if (!entity->unk58 || (entity->unk58--, unk34 & 0x100000)) {
-                    if (!D_800973FC || unk34 & 0x2100 ||
-                        (unk34 & 0x200 && !(D_8003C8C4 & 3))) {
-                        g_CurrentEntity = entity;
-                        entity->pfnUpdate(entity);
-                        entity->unk44 = 0;
-                        entity->hitFlags = 0;
-                    }
-                }
-            }
-        } else {
-            g_CurrentEntity = entity;
-            entity->pfnUpdate(entity);
-            entity->unk44 = 0;
-            entity->hitFlags = 0;
-        }
-    }
-}
+#include "../update.h"
 
 void func_80188514(void) {
     Entity* entity;
@@ -1771,8 +1687,9 @@ void func_80188514(void) {
             if (!(entity->flags & FLAG_UNK_10000))
                 continue;
             if (entity->flags & 0xF) {
-                entity->palette = D_80180690[entity->nFramesInvincibility << 1 |
-                                             LOH(entity->flags) & 1];
+                entity->palette =
+                    UNK_Invincibility0[entity->nFramesInvincibility << 1 |
+                                       LOH(entity->flags) & 1];
                 entity->flags--;
                 if ((entity->flags & 0xF) == 0) {
                     entity->palette = entity->unk6A;
